修正+リファクタリング: クラス名の変更とリファクタリング
